Update README.md
# IPv6
HackathonIPv6
Tunnel Vision
This project is to make sure the packet reaches the destination on a shortest path. But if its travelling through a tunnel we may don’t know there can be another best path out of the tunnel if the packet exits the tunnel before the end of the tunnel sometimes it may be able to select that path. But the reality packet will end when the tunnel finish. It cannot exit from the tunnel. If the source can identify the path where can it exit may help for this. But how do you know that your packet is going through a tunnel? our research is to find whether the packet using a tunnel from source to destination.

Team Members
Shamiya Meezan 
Ziyam Abdeen 
Harinarayan K
Santhosh Kumar
Yogeshwaran K
Description
The plan is to create a tunnel Environment and send the packet from Source to Destination gradually keep on increase the size of the packet. If a Packet is going through a tunnel it will add some more headers to packet where the packet size will increase but tunnel will accept 9192 bytes. If more than that it will split the packets into small sizes and send to the destination via tunnel. If we can capture the destination packets if the size split from the origin.
Difficulties
No one done a such project like this. we were very short of information this regard but we found the mechanism how we can achieve this. More support from community can take this to the next level.
Progress
The project was not completed within the time limits. We need more supports from some programmers to develop a tool for this as we can’t do all these test by manually.